Bug 2974: Fixed a bug when renaming the index directory.
git-svn-id: https://svn.imcode.com/imcode/imcms/branches/2-BRANCH@4366 bd66a97b-2aff-0310-9095-89ca5cabf5a6

+++ server/src/imcode/util/FileUtility.java
@@ -96,12 +96,16 @@ public class FileUtility {
     public static void backupRename(File source, File destination) throws IOException {
         if (!source.equals(destination)) {
             File oldDestination = new File(destination.getParentFile(), destination.getName() + ".old");
-            FileUtils.forceDelete(oldDestination);
+            if (oldDestination.exists()) {
+                FileUtils.forceDelete(oldDestination);
+            }
             if (destination.exists()) {
                 fatalRename(destination, oldDestination);
             }
             fatalRename(source, destination);
-            FileUtils.forceDelete(oldDestination);
+            if (oldDestination.exists()) {
+                FileUtils.deleteDirectory(oldDestination);
+            }
         }
     }
